Restructure presentation packages
