+++ QRCodeGenerator/app/src/main/java/com/example/qrcodegenerator/presentation/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.qrcodegenerator
+package com.example.qrcodegenerator.presentation
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.example.qrcodegenerator.ui.theme.QRCodeGeneratorTheme
+import com.example.qrcodegenerator.presentation.core.theme.QRCodeGeneratorTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
